addprod done

+++ src/Component/Pages/products.js
@@ -1,5 +1,6 @@
 import { useEffect, useState } from "react";
 import { Link } from "react-router-dom";
+import Swal from "sweetalert2";
 
 function Products() {
   const [Products, setProducts] = useState([]);
@@ -11,11 +12,25 @@ function Products() {
   useEffect(() => {
     GetAllProd();
   }, []);
-
+  
+  // fUNC TO DELETE PRODUCTS
   const deProductID = (PID) => {
-    fetch(`http://localhost:9000/products/${PID}`, { method: "delete" })
-      .then((res) => res.json())
-      .then((data) => GetAllProd());
+    Swal.fire({
+      title: `Are you sure u want delet ID ${PID}`,
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire("Deleted!", `Your ID ${PID} has been deleted.`, "success");
+        fetch(`http://localhost:9000/products/${PID}`, { method: "delete" })
+          .then((res) => res.json())
+          .then((data) => GetAllProd());
+      }
+    });
   };
   return (
     <>
